fix: fix web-component tests
