+++ package.json
@@ -33,15 +33,10 @@
   "devDependencies": {
     "@commitlint/cli": "^19.0.0",
     "@commitlint/config-conventional": "^19.0.0",
-    "@open-wc/testing": "^4.0.0",
-    "@playwright/test": "^1.39.0",
     "@types/fs-extra": "^11.0.4",
     "@types/mocha": "^10.0.3",
     "@types/react": "^18.2.36",
     "@types/sinon": "^17.0.0",
-    "@web/test-runner": "^0.19.0",
-    "@web/test-runner-commands": "^0.9.0",
-    "@web/test-runner-playwright": "^0.11.0",
     "babel-preset-carbon": "^0.0.14",
     "chalk": "^5.4.1",
     "change-case": "^5.4.4",
